+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+AWS secrets connector simplified
+
+### Added
+- Source event injection into the AWS lambda
 
 ## 3.0.0-beta.0 - 2024-06-16
 AWS secrets connector added